feat: alert 페이지에서 click, hovering 추가
