+++ src/router/projectRouter.js
@@ -4,6 +4,8 @@ import projectCreate from "@/views/project/ProjectCreate2.vue"
 import ProjectEdit2 from "@/views/project/ProjectEdit2.vue"
 import ProjectList from "@/views/project/ProjectList.vue"
 import ProjectView from "@/views/project/ProjectView3.vue"
+import ChatroomView from '@/views/member/ChatroomView.vue';
+
 export const projectRouter = [
     {
         path: '/editorComponent',
@@ -29,6 +31,11 @@ export const projectRouter = [
         path: '/project/:projectId',
         name: 'ProjectView',
         component: ProjectView
+    },
+    {
+        path: '/project/:projectId/chatroom/:chatroomId',
+        name: 'ChatroomView',
+        component: ChatroomView,
     },
 
 ]
